requirements.txt
feedparser
yfinance
google-generativeai
edge-tts
